backend/requirements.txt
Django==2.2.19
psycopg2-binary==2.9.3
Pillow==9.0.1
djangorestframework==3.13.1
djoser==2.1.0
drf-extra-fields==3.4.0
django-filter==21.1